requirements.txt
atomicwrites==1.4.0
certifi==2021.10.8
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
gunicorn==20.1.0
importlib-metadata==4.8.2
importlib-resources==5.2.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.2
Mako==1.1.4
MarkupSafe==2.0.1
networkx==2.5
olefile==0.46
pluggy==0.13.1
ply==3.11
psycopg2==2.8.6
py==1.9.0
python-dateutil==2.8.2
python-dotenv==0.19.2
six==1.16.0
smop==0.41
SQLAlchemy==1.4.27
toml==0.10.2
Werkzeug==2.0.2
wincertstore==0.2
zipp==3.6.0
numpy==1.21.2
yfinance==0.1.63
pandas==1.3.3
plotly==5.4.0
